inventory_view: add caching per user and access level
